Finish majority of layout and styling.
Completed almost the entire website (forgot to make commits before).
The resume seection is still unimplemented as well as the about text.
Dark mode implemented, and majority of the styling of the different
components. Will probalby not consider dynamically fecthing projectData
from a DB, since the number of projects currently is very low.